migrate to clap beta
