+++ Cargo.toml
@@ -21,7 +21,7 @@ atomic = "0.4"
 atty = "0.2"
 bitflags = "1.2"
 chrono = "0.4"
-clap = "2"
+clap = "=3.0.0-beta.2"
 # TODO: Update this
 dbus = "0.6"
 digest = "0.9"
@@ -61,7 +61,6 @@ serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
 sha2 = "0.9"
-shrinkwraprs = "0.3"
 sys-mount = "1"
 sysfs-class = { git = "https://github.com/pop-os/sysfs-class" }
 systemd-boot-conf = "0.2"
@@ -74,7 +73,6 @@ cascade = "1.0.0"
 async-process = "=0.1.1"
 smolscale = "0.1.0"
 users = "0.11.0"
-structopt = "0.3.21"
 
 [patch.crates-io]
 sysfs-class = { git = "https://github.com/pop-os/sysfs-class" }
