+++ src/cli/daemon.rs
@@ -1,12 +1,12 @@
-use crate::Error;
-use structopt::StructOpt;
+use clap::Clap;
+use pop_upgrade::daemon::DaemonError;
 
 /// launch a daemon for integration with control centers like GNOME's
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Clap)]
 pub struct Daemon {}
 
 impl Daemon {
-    pub fn run(&self) -> Result<(), Error> {
+    pub fn run(&self) -> Result<(), DaemonError> {
         pop_upgrade::daemon::Daemon::init()?;
         Ok(())
     }
