+++ src/cli/release/dismiss.rs
@@ -1,14 +1,7 @@
-use crate::{notify::notify, Error};
-use chrono::{offset::TimeZone, Utc};
 use pop_upgrade::{
-    client::{Client, Error as ClientError, ReleaseInfo},
-    daemon::{DismissEvent, DISMISSED, INSTALL_DATE},
-    misc,
-    release::eol::{EolDate, EolStatus},
+    client::{Client, Error as ClientError},
+    daemon::DismissEvent,
 };
-use std::{convert::TryFrom, fs, path::Path};
-use structopt::StructOpt;
-use ubuntu_version::{Codename, Version as UbuntuVersion};
 
 pub fn run(client: &Client) -> Result<(), ClientError> {
     let devel = pop_upgrade::development_releases_enabled();
